+++ Changelog.md
@@ -1,3 +1,18 @@
+# 2 - Sending multiple messages - [10/01/2023]
+
+- [] Send multiple apps and more than 10 news at once
+
+**Changed Files**
+
+        modified:   .gitignore
+        modified:   Changelog.md
+        new file:   src/discordcommunication/Services/PayloadMappingService.py
+        new file:   src/discordcommunication/Utils/configuration.py
+        modified:   src/discordcommunication/discordServer.py
+        modified:   src/requirements.txt
+
+---
+
 # 1 - First Commit - [04/01/2023]
 
 - [x] Initial Structure